+++ packages/lan/app/routeHandlers/labs/getLabRequestList.js
@@ -3,7 +3,7 @@ import { VISIBILITY_STATUSES } from 'shared/constants/importable';
 
 import { NOTE_RECORD_TYPES } from 'shared/constants/notes';
 
-import { getResourceList } from '../../routes/apiv1/crudHelpers';
+import { getResourceList } from 'shared/utils/crudHelpers';
 
 export const getLabRequestList = (foreignKey = '', options = {}) =>
   asyncHandler(async (req, res) => {
