+++ packages/lan/app/routeHandlers/patients/activeCovid19PatientsHandler.js
@@ -1,7 +1,6 @@
 import { QueryTypes } from 'sequelize';
-import { arrayToDbString } from 'shared/utils';
+import { arrayToDbString, renameObjectKeys } from 'shared/utils';
 import { DIAGNOSIS_CERTAINTY, ENCOUNTER_TYPES } from 'shared/constants';
-import { renameObjectKeys } from '../../utils/renameObjectKeys';
 import { makeFilter } from '../../utils/query';
 import { createPatientFilters } from '../../utils/patientFilters';
