+++ packages/lan/app/routes/apiv1/appointments.js
@@ -2,7 +2,7 @@ import express from 'express';
 import asyncHandler from 'express-async-handler';
 import { startOfDay } from 'date-fns';
 import { Op, Sequelize } from 'sequelize';
-import { simplePost, simplePut } from './crudHelpers';
+import { simplePost, simplePut } from 'shared/utils/crudHelpers';
 import { escapePatternWildcard } from '../../utils/query';
 
 export const appointments = express.Router();
