+++ packages/lan/app/routes/apiv1/certificateNotification.js
@@ -1,6 +1,6 @@
 import express from 'express';
 
-import { simplePost } from './crudHelpers';
+import { simplePost } from 'shared/utils/crudHelpers';
 
 export const certificateNotification = express.Router();
