+++ packages/lan/app/routes/apiv1/documentMetadata.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 
-import { paginatedGetList } from './crudHelpers';
+import { paginatedGetList } from 'shared/utils/crudHelpers';
 
 export const documentMetadata = express.Router();
