+++ packages/lan/app/routes/apiv1/encounter.js
@@ -10,9 +10,6 @@ import {
   VITALS_DATA_ELEMENT_IDS,
   IMAGING_REQUEST_STATUS_TYPES,
 } from 'shared/constants';
-import { uploadAttachment } from '../../utils/uploadAttachment';
-import { notePageListHandler } from '../../routeHandlers';
-
 import {
   simpleGet,
   simpleGetHasOne,
@@ -21,7 +18,10 @@ import {
   permissionCheckingRouter,
   runPaginatedQuery,
   paginatedGetList,
-} from './crudHelpers';
+} from 'shared/utils/crudHelpers';
+import { uploadAttachment } from '../../utils/uploadAttachment';
+import { notePageListHandler } from '../../routeHandlers';
+
 import { getLabRequestList } from '../../routeHandlers/labs';
 
 export const encounter = express.Router();
