+++ packages/lan/app/routes/apiv1/invoice/invoiceLineItems.js
@@ -2,8 +2,13 @@ import asyncHandler from 'express-async-handler';
 import { Op } from 'sequelize';
 import { NotFoundError } from 'shared/errors';
 import { INVOICE_LINE_ITEM_STATUSES } from 'shared/constants';
-import { permissionCheckingRouter, simpleGet, simpleGetList, simplePut } from '../crudHelpers';
-import { renameObjectKeys } from '../../../utils/renameObjectKeys';
+import {
+  permissionCheckingRouter,
+  simpleGet,
+  simpleGetList,
+  simplePut,
+} from 'shared/utils/crudHelpers';
+import { renameObjectKeys } from 'shared/utils';
 import { getPotentialInvoiceLineItems } from './getPotentialInvoiceLineItems';
 
 export const invoiceLineItemsRoute = permissionCheckingRouter('read', 'Invoice');
