+++ packages/lan/app/routes/apiv1/invoice/invoices.js
@@ -3,7 +3,7 @@ import asyncHandler from 'express-async-handler';
 import { customAlphabet } from 'nanoid';
 import { NotFoundError, InvalidParameterError } from 'shared/errors';
 import { INVOICE_STATUSES, INVOICE_PAYMENT_STATUSES } from 'shared/constants';
-import { simplePut } from '../crudHelpers';
+import { simplePut } from 'shared/utils/crudHelpers';
 
 import { invoiceLineItemsRoute } from './invoiceLineItems';
 import { invoicePriceChangeItemsRoute } from './invoicePriceChangeItems';
