+++ packages/lan/app/routes/apiv1/patient/patientFieldDefinition.js
@@ -2,7 +2,7 @@ import asyncHandler from 'express-async-handler';
 import { QueryTypes } from 'sequelize';
 
 import { HIDDEN_VISIBILITY_STATUSES } from 'shared/constants/importable';
-import { permissionCheckingRouter } from '../crudHelpers';
+import { permissionCheckingRouter } from 'shared/utils/crudHelpers';
 
 export const patientFieldDefinition = permissionCheckingRouter('read', 'Patient');
