+++ packages/lan/app/routes/apiv1/patient/patientRelations.js
@@ -4,7 +4,11 @@ import { QueryTypes, Sequelize } from 'sequelize';
 import { getPatientAdditionalData } from 'shared/utils';
 import { HIDDEN_VISIBILITY_STATUSES } from 'shared/constants/importable';
 
-import { simpleGetList, permissionCheckingRouter, runPaginatedQuery } from '../crudHelpers';
+import {
+  simpleGetList,
+  permissionCheckingRouter,
+  runPaginatedQuery,
+} from 'shared/utils/crudHelpers';
 import { patientSecondaryIdRoutes } from './patientSecondaryId';
 import { patientDeath } from './patientDeath';
 import { patientProfilePicture } from './patientProfilePicture';
