+++ packages/lan/app/routes/apiv1/program.js
@@ -1,7 +1,12 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 import { getFilteredListByPermission } from 'shared/utils/getFilteredListByPermission';
-import { simpleGet, simplePut, simplePost, permissionCheckingRouter } from './crudHelpers';
+import {
+  simpleGet,
+  simplePut,
+  simplePost,
+  permissionCheckingRouter,
+} from 'shared/utils/crudHelpers';
 
 export const program = express.Router();
