+++ packages/lan/app/routes/apiv1/referenceData.js
@@ -1,6 +1,6 @@
 import express from 'express';
 
-import { simpleGet, simplePut, simplePost } from './crudHelpers';
+import { simpleGet, simplePut, simplePost } from 'shared/utils/crudHelpers';
 
 export const referenceData = express.Router();
