+++ packages/shared-src/package.json
@@ -40,6 +40,8 @@
     "date-fns": "^2.19.0",
     "date-fns-tz": "^1.3.6",
     "eslint": "^5.15.2",
+    "express": "^4.16.2",
+    "express-async-handler": "^1.1.4",
     "jsonwebtoken": "^8.4.0",
     "libhoney": "^3.1.1",
     "lodash": "^4.17.11",
